requirements.txt
Django==2.0.6
django-ratelimit==1.1.0
Pillow==5.1.0
pytz==2018.4
